% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{all_genes}
\alias{all_genes}
\title{Gene Expression Analysis Dataset}
\format{
A data frame with the following columns:
\describe{
  \item{genes}{Character. Gene name or identifier.}
  \item{baseMean}{Numeric. The base mean value for the gene across samples.}
  \item{log2FoldChange}{Numeric. The log2 fold change of gene expression. Positive values indicate upregulation and negative values indicate downregulation.}
  \item{lfcSE}{Numeric. Standard error of the log2 fold change.}
  \item{stat}{Numeric. The Wald statistic for the gene's expression change.}
  \item{pvalue}{Numeric. Raw p-value for the test of the gene's expression change.}
  \item{padj}{Numeric. Adjusted p-value for multiple testing corrections.}
}
}
\usage{
all_genes
}
\description{
A dataset containing gene expression analysis results.
It has seven columns capturing various statistics related to gene expression.
}
\keyword{datasets}
