% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{attention_genes}
\alias{attention_genes}
\title{Attention Genes Dataset}
\format{
A data frame with 10 rows and 7 variables:
\describe{
  \item{genes}{Character. Gene name or identifier.}
  \item{baseMean}{Numeric vector: Base mean expression level of genes}
  \item{log2FoldChange}{Numeric vector: Log2 Fold Change of gene expression}
  \item{lfcSE}{Numeric vector: Standard error for log2 fold change}
  \item{stat}{Numeric vector: Wald statistic for the gene's expression change}
  \item{pvalue}{Numeric vector: Raw p-value from Wald test}
  \item{padj}{Numeric vector: Adjusted p-value for multiple testing using the Benjamini-Hochberg procedure}
}
}
\usage{
attention_genes
}
\description{
A dataset containing specific genes of interest referred to as "attention genes".
}
\keyword{datasets}
