% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_de_source}
\alias{detect_de_source}
\title{Detect DE pipeline from column names}
\usage{
detect_de_source(cols)
}
\arguments{
\item{cols}{Character vector of column names.}
}
\value{
One of \code{"deseq2"}, \code{"edger"}, \code{"limma"}, or
  errors if unrecognized.
}
\description{
Detect DE pipeline from column names
}
\keyword{internal}
