% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ggvolc-onAttach}
\alias{ggvolc-onAttach}
\title{Print Welcome Message and ASCII Art upon Loading `ggvolc`}
\description{
This function is executed when the `ggvolc` package is attached to the R session.
It prints a welcome message and an ASCII representation related to the package.
}
\examples{
# The function is automatically called when you use:
# library(ggvolc)
}
\seealso{
\code{\link[base]{library}}
}
\keyword{internal}
