% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvolc.R
\name{ggvolc}
\alias{ggvolc}
\title{Create a Volcano Plot}
\usage{
ggvolc(
  data1,
  data2 = NULL,
  size_var = NULL,
  p_value = 0.05,
  fc = 1,
  not_sig_color = "grey82",
  down_reg_color = "#00798c",
  up_reg_color = "#d1495b",
  add_seg = FALSE
)
}
\arguments{
\item{data1}{Data frame for the primary dataset.}

\item{data2}{Data frame for the secondary dataset. Default is NULL.}

\item{size_var}{Variable for determining the size of points. Options are "log2FoldChange" and "pvalue". Default is "log2FoldChange".}

\item{p_value}{Threshold for statistical significance. Default is 0.05.}

\item{fc}{Fold change threshold for determining upregulated or downregulated genes. Default is 1.}

\item{not_sig_color}{Color for non-significant genes. Default is "grey82".}

\item{down_reg_color}{Color for downregulated genes. Default is "#00798c".}

\item{up_reg_color}{Color for upregulated genes. Default is "#d1495b".}

\item{add_seg}{Logical. If TRUE, dashed lines will be added to the plot indicating the p-value and fold change thresholds. Default is FALSE.}
}
\value{
A ggplot2 object displaying the volcano plot.
}
\description{
This function creates a volcano plot using ggplot2 based on provided datasets.
It is particularly useful for visualizing differential gene expression data.
}
\examples{
\dontrun{
# Assuming df1 and df2 are your data frames:
library(ggvolc)
ggvolc(df1, df2)
}

}
